Split sound code to new file on client
